requirements.txt
# Web相关
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# 配置和数据处理
PyYAML==6.0.1
python-dotenv==1.0.0

# 调度和时间处理
schedule==1.2.0
pytz==2023.3

# arXiv相关
arxiv==2.1.0
feedparser==6.0.10

# 邮件处理
email-validator==2.1.0

# PDF处理
PyPDF2==3.0.1
pymupdf>=1.24.0  # arxiv_digest: PDF text extraction (fitz)

# OpenAI API (支持OpenAI兼容接口)
openai==1.109.1

# ModelScope（模型下载）
modelscope

# 本地模型推理服务（vLLM，提供 OpenAI 兼容 API）
# 需 GPU 与较新驱动，建议在独立环境中安装
vllm